size_t instead of int
git-svn-id: file:///home/roberto/Git/svn@8996 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 0658a97d9bed34716680ae76052372258525a140

+++ trunk/xmipp/libraries/data/rwIMAGIC.cpp
@@ -203,7 +203,7 @@ int  ImageBase::readIMAGIC(size_t select_img)
     MD.clear();
     MD.resize(_nDim,MDL::emptyHeader);
     double daux=1.;
-    for (int i = 0; i < _nDim; ++i )
+    for ( size_t i = 0; i < _nDim; ++i )
     {
         if ( fread( header, IMAGICSIZE, 1, fhed ) < 1 )
             return(-2);
